added counter observer

+++ Exam/Include/Reaction/Agent.h
@@ -9,13 +9,13 @@
 #include <vector>
 
 class Agent{
-    int value;
+    size_t value;
     std::string id;
 public:
     void setValue(int v){value = v;}
     [[nodiscard]] int getValue() const{return value;};
     [[nodiscard]] std::string getId() const{return id;};
-    Agent(std::string id, const int& val): value(val), id(std::move(id)){};
+    Agent(std::string id, const size_t& val): value(val), id(std::move(id)){};
     std::string toString() const;
 
     //REQUIREMENT 2a - Pretty printer in human readable format
